requirements.txt
streamlit
cohere
groq
python-dotenv
PyPDF2
python-docx
pandas
openpyxl